+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 ## 0.0.39
 - [commands-spigot] Added `WorldParser` to SpigotParsers
+- [commands-spigot] Fixes `WorldParser` to support on legacy versions
 - [common] Bump Minecraft Version 1.21.2 - 1.21.5
+- [commands] Add Contextual Report for "Not Found" Results in `MapBasedParser` as `NoSuchElementInMapResponse`
 
 ## 0.0.38
 - [commands] Add `WrappedArgumentParser` for better handling of argument parsers